int conversion

+++ src/models.py
@@ -982,7 +982,7 @@ def estimate_hypers(step, problem):
 
     for v, i in pc.hyperparameters.iteritems():
         d = mtrace.get_values(
-            v, combine=True, burn=(n_steps / 2.), thin=2., squeeze=True)
+            v, combine=True, burn=int(n_steps / 2.), thin=2, squeeze=True)
         lower = d.min(axis=0)
         upper = d.max(axis=0)
         logger.info('Updating hyperparameter %s from %f, %f to %f, %f' % (
